misc: fs_loader: reorganize and split to FS loader and FW UCLASS
In preparation to the introduction of variant of the FS loader,
reorganize and split the driver to generic fw_loader UCLASS and
specific fs_loader function. Create a dedicated directory for the loader
and move the internal structs and functions to a dedicated header file.
While at it also drop some redundant check and rename some variables as we
are updating the code for the UCLASS rework.

This will permit to reuse all the property and logic of FS loader
with container that are not exactly a readable filesystem.
